Add files via upload
Attached Presentation shows the techniques used to solve this Classification problem using Different Classification techniques. The operations are done using Azure ML Platform. Text analysis is also done for the column description that shows the Top 10 Entities used and language detection model for the same.
For further references, link for the azure gallery is also attached to see and tune the accuracy of used models.